AU-specific reports and fields
A major change: dev/production info form now separated out from main course info and expanded to include fields the may be somewhat specific to Athabasca University but most of which would have value elsewhere. Some interface tweaks and rationalizations as a result. A number of new report types matching those needed for the AU course development process. This release is still a bit glitchy in places and some tidying is needed here and there, but it works well enough to do the job